Refactor external service interfaces and implementations to improve package structure
