+++ src/main/java/Application/MarketService.java
@@ -3,9 +3,9 @@ package Application;
 import Domain.IMarketFacade;
 import Domain.User.IUser;
 import Domain.User.IUserFacade;
-import Infrastructure.INotificationService;
-import Infrastructure.IPaymentService;
-import Infrastructure.ISupplyService;
+import Domain.ExternalServices.INotificationService;
+import Domain.ExternalServices.IPaymentService;
+import Domain.ExternalServices.ISupplyService;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.Date;
