+++ src/main/java/Domain/ExternalServices/INotificationService.java
@@ -1,4 +1,4 @@
-package Infrastructure;
+package Domain.ExternalServices;
 
 public interface INotificationService {
