+++ src/main/java/Domain/IMarketFacade.java
@@ -5,9 +5,9 @@ import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import Infrastructure.IPaymentService;   
-import Infrastructure.INotificationService;
-import Infrastructure.ISupplyService;
+import Domain.ExternalServices.INotificationService;
+import Domain.ExternalServices.IPaymentService;
+import Domain.ExternalServices.ISupplyService;
 import Domain.User.IUserFacade;
 import Domain.User.IUser;
 import Domain.Store.IStoreFacade;
